requirements.txt
Flask==2.1.0
Flask-MySQL==1.5.2
mysql-connector-python==8.0.25
python-dotenv==0.19.0
